Assign menu_state_t::d only on entering a menu
As opposed to doing this on initializing one how it was done before.
Previous approach is not just weird, it's problematic as it loses
information about the previous menu while preparing the new one but
before actually switching to it.  This is important for making switching
between menus easier.

undolist_menu.c was updated because menus_set_pos() should not be used
during menu setup (only in handlers), instead m.pos must be set directly.